Fix the cdn path of the images for Amazon s3
At the moment the cdn replace system change images path from
'..dist/img/'
to
'BUCKETURI/BUCKETNAME/SUBFOLDER/dist/img/'

That however return the error "The bucket you are attempting to access must be addressed using the specified endpoint. Please send all future requests to this endpoint.".

To solve the issue, you should change set as BUCKETURI the endpoint of the bucket (ex. 'http://BUCKETNAME.s3-website-eu-west-1.amazonaws.com') and remove the BUCKETNAME after the it in the replacing task.